requirements.txt
fastapi==0.99.1
redis[hiredis]
uvicorn==0.20.0
python-dotenv==1.0.0